Вынес в отдельный модуль render.js  функции renderElement, createElement, перечисление REnderPosition. Переписал функцию renderElement, чтобы она принимала на вход компонент, а не дом-элемент
