+++ src/components/card.js
@@ -1,5 +1,5 @@
 import {MonthNames} from '../constants.js';
-import {formatTime} from '../utils.js';
+import {formatTime} from '../utils/utils.js';
 import AbstractComponent from './abstract-component.js';
 
 const createHashtagsMarkup = (hashtags) => {
